# Conflicts:
#	.env
#	app/Http/Controllers/PageController.php
#	routes/web.php

+++ database/migrations/2017_04_13_131057_update_users_table.php
@@ -25,8 +25,10 @@ class UpdateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('is_admin');
-        });
+        if (Schema::hasColumn('users', 'is_admin')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('is_admin');
+            });
+        }
     }
 }
